Various Modifications, check the description.
Refactored the code (introduced more procedures, eliminated unused macros, labels and some comments), minor changes to the checkShip functions, dynamicBall function, introduced wave motion, and wave collision and wave slow reduction, imrpoved the speed of the ball for level two, and the two levels are now seperate.
A bug that I'm currently trying to solve is that sometime* when the left ship wins the end-game screen does not redirect to the main menu.